requirements.txt
discord.py
pymongo.py
dnspython.py